refactor: split up codes for modularization
