+++ src/neosearch/app/settings.py
@@ -1,11 +1,4 @@
-from contextlib import asynccontextmanager
-from fastapi import FastAPI
-from fastapi.middleware.gzip import GZipMiddleware
-from uvicorn.middleware.proxy_headers import ProxyHeadersMiddleware
 import os
-from functools import cache
-from pathlib import Path
-import toml
 
 from llama_index.llms.openai import OpenAI
 from llama_index.llms import Ollama
@@ -13,15 +6,11 @@ from llama_index.embeddings.openai import OpenAIEmbedding
 from llama_index.core.settings import Settings
 
 # custom module
-from neosearch.app.middlewares import RequestLogger, RequestID
-from neosearch.app.utils.logging import Logger
-from neosearch.app.utils.gc_tuning import gc_optimization_on_startup
 from neosearch.app.utils.configs import Configs
 
-logger = Logger()
 config = Configs()
 
-def init_settings():
+def init_llm_settings() -> None:
     llm_config = config.get_llm_configs()
     llm_type = llm_config.get("type", "openai")
 
@@ -34,63 +23,13 @@ def init_settings():
     else:
         raise ValueError(f"Invalid LLM type: {llm_type}")
 
+
+def init_settings() -> None:
+    init_llm_settings()
+
     Settings.embed_model = OpenAIEmbedding(
-        model="text-embedding-3-small", embed_batch_size=100
+        model="text-embedding-3-small",
+        embed_batch_size=100
     )
     Settings.chunk_size = 1024
     Settings.chunk_overlap = 20
-
-
-@cache
-def project_root() -> Path:
-    """Find the project root directory by locating pyproject.toml."""
-    base_dir = Path(__file__).parent
-
-    for parent_directory in base_dir.parents:
-        if (parent_directory / "pyproject.toml").is_file():
-            return parent_directory
-    raise FileNotFoundError("Could not find project root containing pyproject.toml")
-
-
-def get_version_from_pyproject_toml() -> str:
-    try:
-        # Probably this is the pyproject.toml of a development install
-        path_to_pyproject_toml = project_root() / "pyproject.toml"
-    except FileNotFoundError:
-        # Probably not a development install
-        path_to_pyproject_toml = None
-
-    if path_to_pyproject_toml is not None:
-        pyproject_version = toml.load(path_to_pyproject_toml)["tool"]["poetry"]["version"]
-        return pyproject_version
-    else:
-        return os.getenv("VERSION", "x.x.x")
-
-
-@asynccontextmanager
-async def lifespan(app: FastAPI):
-    logger.get_logger()
-    # gc optimization
-    gc_optimization_on_startup()
-
-    #TODO open redis connection for lifespan
-    yield
-
-    #TODO: Add code to clean up the app context
-
-
-def init_app() -> FastAPI:
-    _version = get_version_from_pyproject_toml()
-    app = FastAPI(title="NeoSearch", version=_version, lifespan=lifespan)
-
-    # add middlewares
-    app.add_middleware(
-        ProxyHeadersMiddleware, trusted_hosts="*"
-    )  # add proxy headers to prevent logging IP address of the proxy server instead of the client
-    app.add_middleware(GZipMiddleware, minimum_size=500)  # add gzip compression
-
-    # add custom middlewares
-    app.add_middleware(RequestLogger)
-    app.add_middleware(RequestID)
-
-    return app
